# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/kitchen/orders/{id}/status": {
            "patch": {
                "description": "Update order status (preparing, ready, served, completed)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kitchen"
                ],
                "summary": "Update order status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New order status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateOrderStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/receipts/scan": {
            "post": {
                "description": "Extract item details from receipt image using AI vision client",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Receipts"
                ],
                "summary": "Scan receipt image (AI Vision)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Receipt Image file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ReceiptAudit"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.OCRItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "domain.OrderStatus": {
            "type": "string",
            "enum": [
                "UNPAID",
                "PENDING",
                "PREPARING",
                "READY",
                "SERVED",
                "COMPLETED",
                "CANCELLED",
                "EXPIRED",
                "REFUNDED",
                "PENDING",
                "COMPLETED",
                "CANCELLED",
                "REFUNDED"
            ],
            "x-enum-varnames": [
                "OrderStatusUnpaid",
                "OrderStatusPending",
                "OrderStatusPreparing",
                "OrderStatusReady",
                "OrderStatusServed",
                "OrderStatusCompleted",
                "OrderStatusCancelled",
                "OrderStatusExpired",
                "OrderStatusRefunded",
                "StatusPending",
                "StatusCompleted",
                "StatusCancelled",
                "StatusRefunded"
            ]
        },
...
        "domain.ReceiptAudit": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "merchant_name": {
                    "type": "string"
                },
                "raw_ocr_json": {
                    "$ref": "#/definitions/domain.ReceiptOCRResult"
                },
                "receipt_date": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
        "domain.ReceiptOCRResult": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.OCRItem"
                    }
                },
                "merchant_name": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
        "dto.UpdateOrderStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "enum": [
                        "PENDING",
                        "PREPARING",
                        "READY",
                        "SERVED",
                        "COMPLETED",
                        "CANCELLED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.OrderStatus"
                        }
                    ]
                }
            }
        },
...
    }
}